Move generic helper to test-suites
